Removed old ProfileFragment and added ProfileActivity
