+++ app/src/main/AndroidManifest.xml
@@ -9,9 +9,11 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
+        <activity android:name=".ui.profile.ProfileActivity"></activity>
         <activity android:name=".ui.LoginActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
@@ -19,7 +21,7 @@
             android:name=".ui.HomeActivity"
             android:windowSoftInputMode="adjustPan">
             <intent-filter>
-                <action android:name="android.intent.action.VIEW"/>
+                <action android:name="android.intent.action.VIEW" />
             </intent-filter>
         </activity>
         <activity
@@ -33,6 +35,7 @@
         <activity
             android:name=".ui.pubs.PubActivity"
             android:theme="@style/AppTheme" />
-        <activity android:name=".ui.SignUpActivity"/>
+        <activity android:name=".ui.SignUpActivity" />
     </application>
+
 </manifest>
