--- app/src/main/java/com/frisky/icebreaker/ui/profile/ProfileFragment.java
@@ -1,44 +1,0 @@
-package com.frisky.icebreaker.ui.profile;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v4.view.ViewPager;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.ImageButton;
-
-import com.frisky.icebreaker.R;
-
-public class ProfileFragment extends Fragment {
-
-    ImageButton mEditButton;
-    ViewPager mProfileImagePager;
-    ProfileImageAdapter mProfileImageAdapter;
-
-    @Nullable
-    @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment_profile_view, null);
-
-        mEditButton = view.findViewById(R.id.button_edit);
-        mEditButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                //TODO Start edit profile activity here
-                Intent editProfile = new Intent(getContext(), EditProfileActivity.class);
-                startActivity(editProfile);
-            }
-        });
-
-        mProfileImageAdapter = new ProfileImageAdapter(getContext());
-
-        mProfileImagePager = view.findViewById(R.id.pager_profile_images);
-        mProfileImagePager.setAdapter(mProfileImageAdapter);
-
-        return view;
-    }
-}
