+++ app/src/main/java/com/frisky/icebreaker/ui/profile/ProfileImageAdapter.java
@@ -16,9 +16,7 @@ public class ProfileImageAdapter extends PagerAdapter {
     LayoutInflater mLayoutInflater;
 
     int[] mImageList = {
-            R.drawable.sample_1,
             R.drawable.sample_2,
-            R.drawable.sample_3,
             R.drawable.sample_4,
             R.drawable.sample_5
     };
