+++ app/src/main/res/layout/content_view_profile.xml
@@ -8,8 +8,4 @@
     tools:context=".ui.profile.ViewProfileActivity"
     tools:showIn="@layout/activity_view_profile">
 
-    <include
-        layout="@layout/fragment_profile_view"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content" />
 </android.support.constraint.ConstraintLayout>
